Assembling stuff in dialog master
